+++ src/test/kotlin/utils/primes/PrimeSieveTest.kt
@@ -4,6 +4,7 @@ import junit.framework.Assert.assertEquals
 import org.junit.Test
 import java.io.File
 import java.lang.Integer.parseInt
+import java.lang.Long.parseLong
 
 class PrimeSieveTest {
     @Test
@@ -19,22 +20,22 @@ class PrimeSieveTest {
 
     @Test
     fun `if maxPrime is 2, it returns the list (2)`() {
-        assertEquals(listOf(2), PrimeSieve(maxPrime = 2).sieve())
+        assertEquals(listOf(2L), PrimeSieve(maxPrime = 2).sieve())
     }
 
     @Test
     fun `if maxPrime is 3, it returns the list (2, 3)`() {
-        assertEquals(listOf(2, 3), PrimeSieve(maxPrime = 3).sieve())
+        assertEquals(listOf(2L, 3L), PrimeSieve(maxPrime = 3).sieve())
     }
 
     @Test
     fun `if maxPrime is 10, it returns the list (2, 3, 5, 7)`() {
-        assertEquals(listOf(2, 3, 5, 7), PrimeSieve(maxPrime = 10).sieve())
+        assertEquals(listOf(2L, 3L, 5L, 7L), PrimeSieve(maxPrime = 10).sieve())
     }
 
     @Test
     fun `if maxPrime is 20, it returns the list (2, 3, 5, 7, 11, 13, 17, 19)`() {
-        assertEquals(listOf(2, 3, 5, 7, 11, 13, 17, 19), PrimeSieve(maxPrime = 20).sieve())
+        assertEquals(listOf(2L, 3L, 5L, 7L, 11L, 13L, 17L, 19L), PrimeSieve(maxPrime = 20).sieve())
     }
 
     @Test
@@ -48,10 +49,10 @@ class PrimeSieveTest {
         assertEquals(knownPrimes, PrimeSieve(maxPrime = 1000_000).sieve())
     }
 
-    private fun loadPrimesList(fileName: String): List<Int> {
+    private fun loadPrimesList(fileName: String): List<Long> {
         val dirPath = File("").canonicalPath
         val actualPath = listOf(dirPath, "src", "test", "kotlin", "utils", "primes", fileName).joinToString(File.separator)
         val reader = File(actualPath).bufferedReader()
-        return reader.readLine().split(",").map { parseInt(it) }
+        return reader.readLine().split(",").map { parseLong(it) }
     }
 }
